% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempawn.R
\name{compute_dummy_95}
\alias{compute_dummy_95}
\title{Compute the 95\% quantile of the target variable sensitivities for the dummy}
\usage{
compute_dummy_95(res_i)
}
\arguments{
\item{res_i}{Results for the sensitivity analysis for the target i}
}
\description{
Compute the 95\% quantile of the target variable sensitivities for the dummy
}
\keyword{internal}
