% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tempawn.R
\name{explore_parameter}
\alias{explore_parameter}
\title{Plot analysis results of temoral PAWN analysis}
\usage{
explore_parameter(
  simulation,
  parameter,
  bins = 10,
  variable = NULL,
  observation = NULL
)
}
\arguments{
\item{tempawn}{The list object that is returned by \code{tempawn()}.}
}
\value{
Returns a ggplot object with the visualization of the simulation ranges
and the individual temoral parameter sensitivities.
}
\description{
Plot analysis results of temoral PAWN analysis
}
