% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempawn.R
\name{finish_progress}
\alias{finish_progress}
\title{Print message for completed process}
\usage{
finish_progress(t0, word)
}
\arguments{
\item{t0}{initial time step}

\item{nmax}{Number of iterations}
}
\description{
Print message for completed process
}
\keyword{internal}
