% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempawn.R
\name{get_sim_stat}
\alias{get_sim_stat}
\title{Compute min/max values for the simulation of each time step}
\usage{
get_sim_stat(sim_tbl)
}
\arguments{
\item{sim_tbl}{Table with simulated time series}
}
\description{
Compute min/max values for the simulation of each time step
}
\keyword{internal}
