% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempawn.R
\name{merge_inp_tgt}
\alias{merge_inp_tgt}
\title{Merge inputs and targets}
\usage{
merge_inp_tgt(inp, tgt)
}
\arguments{
\item{inp}{Input}

\item{tgt}{Target}
}
\description{
Merge inputs and targets
}
\keyword{internal}
