% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempawn.R
\name{pawn_bin}
\alias{pawn_bin}
\title{Compute the PAWN index for the defined bins for one input and one target}
\usage{
pawn_bin(dat, idx, stat)
}
\arguments{
\item{dat}{Table that provides the input and the target pairs}

\item{idx}{index that defines the bin classes}

\item{stat}{The summary statistics function that is implemented to compute
the PAWN index.}
}
\description{
Compute the PAWN index for the defined bins for one input and one target
}
\keyword{internal}
