% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempawn.R
\name{pawn_i}
\alias{pawn_i}
\title{Compute pawn_bin for all inputs and one target}
\usage{
pawn_i(inp_tgt_i, idx, stat)
}
\arguments{
\item{inp_tgt_i}{List that provides the input and the target pairs for all inputs}

\item{idx}{index that defines the bin classes}

\item{stat}{The summary statistics function that is implemented to compute
the PAWN index.}
}
\description{
Compute pawn_bin for all inputs and one target
}
\keyword{internal}
