% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempawn.R
\name{pawn_ij}
\alias{pawn_ij}
\title{Compute pawn_bin for one input and one target}
\usage{
pawn_ij(inp_tgt_ij, idx, stat)
}
\arguments{
\item{inp_tgt_ij}{Table that provides the input and the target pairs}

\item{idx}{index that defines the bin classes}

\item{stat}{The summary statistics function that is implemented to compute
the PAWN index.}
}
\description{
Compute pawn_bin for one input and one target
}
\keyword{internal}
