% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempawn.R
\name{tempawn}
\alias{tempawn}
\title{Calculate temporal PAWN indices}
\usage{
tempawn(sim, var = NULL, stat = median, bins = 25, dummy = TRUE, cores = NULL)
}
\arguments{
\item{sim}{The simulation output object that results from a simulation with
run_swat2012() or run_swatplus() from the SWATplusR package.}

\item{var}{String vector of simulated variables that should be included in the
sensitivity analysis. \code{Default = NULL} means all variables are included.}

\item{stat}{The summary statistics that is used to compute the PAWN sensitivity
index. Pianosi and Wagener (2015) suggest to implement the median or max.
\code{Default = median}}

\item{bins}{Number of bins into which each parameter dimension is separated.
\code{Default = 25}}

\item{dummy}{\code{dummy = TRUE} includes the calculation of the dummy sensitivity.
This option can tremendously increase the calculation time, but provides very
important information. \code{Default = TRUE}}

\item{cores}{Parallel computing is implemented to reduce computation time. By
default all cores are used. If a lower number of cores should be used set cores
to a certain value.}
}
\value{
Returns a list of tables where each table provides the sensitivities
  for each time step of the simulation outputs to all the analyzed model
  parameters.
}
\description{
tempawn calculates the PAWN sensitivity index for each time step of simulation
results that were generated with the run_swat*() functions from the SWATplusR
package. The PAWN method is implemented according to Pianosi and Wagener (2018).
tempawn optionally includes the computation of the temporal sensitivity of a
dummy parameter.
}
\references{
\cite{Pianosi, F. and Wagener T.: Distribution-based sensitivity analysis
      from a generic input-output sample, Environmental Modelling & Software,
      108, 197-207, \url{https://doi.org/10.1016/j.envsoft.2018.07.019}, 2018}


\cite{Pianosi, F. and Wagener T.: A simple and efficient method for global
      sensitivity analysis based on cumulative distribution functions,
      Environmental Modelling & Software, 67, 1-11,
      \url{https://doi.org/10.1016/j.envsoft.2015.01.004}, 2015}
}
